se agrega espacio en función -sprintf- para que funcione correctamente la funcion -split- en HMIcomRTU con el último elemento del buffer

+++ nfc/inc/rtu_com_hmi.h
@@ -152,7 +152,7 @@ typedef struct {
 
 } RTUData_t;
 
-void NetValuesToSendFromRTU(int16_t iServerStatus, RTUData_t *pRTUDataTx);
+void NetValuesToSendFromRTU(int8_t iServerStatus, RTUData_t *pRTUDataTx);
 
 /*	----------------------------------------------------------------------------------------- */
 /*	----------------------------	-	TASK TRIGGER & STATUS HANDLER	-	----------------- */
